backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
stripe==11.2.0
google-generativeai==0.8.6
python-magic==0.4.27
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
alembic==1.14.0
reportlab==4.2.5
PyPDF2==3.0.1
PyMuPDF==1.24.14
python-docx==1.1.2
redis==5.2.0
